first commit
add all the files for the pizza order calculator app that calculate the price of pizzas that will be purchase during a promotioinal period. The number of pizzas will be entered by the PizzasOnly salesperson:
- One large pizza will cost $6.45
- Two large pizza will cost $12.00
- Three large pizza will cost $14.00
Four or more pizzas will use a combination of the above prices to ensure thet best price for the customer, for example, the best price for five (5) pizzas would be two pizzas ($12) + three pizzas ($14).